+++ layoutabi/__init__.py
@@ -13,7 +13,7 @@ from .errors import (
 from .explain import explain
 from .supported import supported
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 
 __all__ = [
     "InvalidArgumentError",